router streams : BalanceProcessor refactoring and unit test
